[xml] Reorganize rules into categories
