+++ docs/pages/pmd/rules/pom.md
@@ -5,9 +5,12 @@ folder: pmd/rules
 ---
 List of rulesets and rules contained in each ruleset.
 
-*   [Basic POM](pmd_rules_pom_basic.html): The Basic POM Ruleset contains a collection of good practices regarding Maven's POM files.
+*   [Codestyle](pmd_rules_pom_codestyle.html): Rules which enforce a specific coding style.
+*   [Errorprone](pmd_rules_pom_errorprone.html): Rules to detect constructs that are either broken, extremely confusing or prone to runtime errors.
 
-## Basic POM
-*   [InvalidDependencyTypes](pmd_rules_pom_basic.html#invaliddependencytypes): While Maven will not failed if you use an invalid type for a dependency in thedependency manageme...
-*   [ProjectVersionAsDependencyVersion](pmd_rules_pom_basic.html#projectversionasdependencyversion): Using that expression in dependency declarations seems like a shortcut, but it can go wrong.By fa...
+## Codestyle
+*   [ProjectVersionAsDependencyVersion](pmd_rules_pom_codestyle.html#projectversionasdependencyversion): Using that expression in dependency declarations seems like a shortcut, but it can go wrong.By fa...
+
+## Errorprone
+*   [InvalidDependencyTypes](pmd_rules_pom_errorprone.html#invaliddependencytypes): While Maven will not failed if you use an invalid type for a dependency in thedependency manageme...
