+++ docs/pages/pmd/rules/xsl.md
@@ -5,9 +5,12 @@ folder: pmd/rules
 ---
 List of rulesets and rules contained in each ruleset.
 
-*   [XPath in XSL](pmd_rules_xsl_xpath.html): This ruleset regroups a collection of good practices regarding XPath querying and functions inside an XSL.
+*   [Codestyle](pmd_rules_xsl_codestyle.html): Rules which enforce a specific coding style.
+*   [Performance](pmd_rules_xsl_performance.html): Rules that flag suboptimal code.
 
-## XPath in XSL
-*   [AvoidAxisNavigation](pmd_rules_xsl_xpath.html#avoidaxisnavigation): Avoid using the 'following' or 'preceeding' axes whenever possible, as these can cutthrough 100% ...
-*   [UseConcatOnce](pmd_rules_xsl_xpath.html#useconcatonce): The XPath concat() functions accepts as many arguments as required so you can have"concat($a,'b',...
+## Codestyle
+*   [UseConcatOnce](pmd_rules_xsl_codestyle.html#useconcatonce): The XPath concat() functions accepts as many arguments as required so you can have"concat($a,'b',...
+
+## Performance
+*   [AvoidAxisNavigation](pmd_rules_xsl_performance.html#avoidaxisnavigation): Avoid using the 'following' or 'preceeding' axes whenever possible, as these can cutthrough 100% ...
