+++ pmd-xml/src/main/resources/category/xml/errorprone.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0"?>
+
+<ruleset name="Errorprone"
+         xmlns="http://pmd.sourceforge.net/ruleset/2.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0 http://pmd.sourceforge.net/ruleset_2_0_0.xsd">
+
+    <description>
+Rules to detect constructs that are either broken, extremely confusing or prone to runtime errors.
+    </description>
+
+    <rule name="MistypedCDATASection"
+          language="xml"
+          since="5.0"
+          message="Potentialy mistyped CDATA section with extra [ at beginning or ] at the end."
+          class="net.sourceforge.pmd.lang.rule.XPathRule"
+          externalInfoUrl="${pmd.website.baseurl}/pmd_rules_xml_errorprone.html#mistypedcdatasection">
+        <description>
+An XML CDATA section begins with a &lt;!CDATA[ marker, which has only one [, and ends with a ]]&gt; marker, which has only two ]. 
+        </description>
+        <priority>3</priority>
+        <properties>
+            <property name="xpath">
+                <value>
+<![CDATA[
+//cdata-section[starts-with(@Image,'[') or ends-with(@Image,']')]
+]]>
+                </value>
+            </property>
+        </properties>
+        <example>
+<![CDATA[
+An extra [ looks like &lt;!CDATA[[]]&gt;, and an extra ] looks like &lt;!CDATA[]]]&gt;.
+]]>
+        </example>
+    </rule>
+
+</ruleset>
