+++ pmd-xml/src/main/resources/category/xsl/documentation.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0"?>
+
+<ruleset name="Documentation"
+         xmlns="http://pmd.sourceforge.net/ruleset/2.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0 http://pmd.sourceforge.net/ruleset_2_0_0.xsd">
+
+    <description>
+Rules that are related to code documentation.
+    </description>
+</ruleset>
