+++ pmd-xml/src/main/resources/rulesets/pom/rulesets.properties
@@ -2,4 +2,17 @@
 # BSD-style license; for more info see http://pmd.sourceforge.net/license.html
 #
 
-rulesets.filenames=rulesets/pom/basic.xml
+rulesets.filenames=\
+    category/pom/codestyle.xml,\
+    category/pom/errorprone.xml
+
+#
+# categories without rules
+#
+# category/pom/bestpractices.xml
+# category/pom/design.xml
+# category/pom/documentation.xml
+# category/pom/multithreading.xml
+# category/pom/performance.xml
+# category/pom/security.xml
+
