+++ pmd-xml/src/main/resources/rulesets/wsdl/rulesets.properties
@@ -3,3 +3,16 @@
 #
 
 rulesets.filenames=
+
+#
+# categories without rules
+#
+# category/wsdl/bestpractices.xml
+# category/wsdl/codestyle.xml
+# category/wsdl/design.xml
+# category/wsdl/documentation.xml
+# category/wsdl/errorprone.xml
+# category/wsdl/multithreading.xml
+# category/wsdl/performance.xml
+# category/wsdl/security.xml
+
