+++ pmd-xml/src/main/resources/rulesets/xml/basic.xml
@@ -8,30 +8,6 @@
 The Basic XML Ruleset contains a collection of good practices which everyone should follow.
   </description>
 
-  <rule name="MistypedCDATASection"
-        language="xml"
-        since="5.0"
-        message="Potentialy mistyped CDATA section with extra [ at beginning or ] at the end."
-        class="net.sourceforge.pmd.lang.rule.XPathRule"
-        externalInfoUrl="${pmd.website.baseurl}/pmd_rules_xml_basic.html#mistypedcdatasection">
-    <description>
-An XML CDATA section begins with a &lt;!CDATA[ marker, which has only one [, and ends with a ]]&gt; marker, which has only two ]. 
-    </description>
-    <priority>3</priority>
-    <properties>
-        <property name="xpath">
-            <value>
-<![CDATA[
-//cdata-section[starts-with(@Image,'[') or ends-with(@Image,']')]
-]]>
-            </value>
-        </property>
-    </properties>
-    <example>
-<![CDATA[
-An extra [ looks like &lt;!CDATA[[]]&gt;, and an extra ] looks like &lt;!CDATA[]]]&gt;.
-]]>
-    </example>
-  </rule>
+    <rule ref="category/xml/errorprone.xml/MistypedCDATASection" deprecated="true" />
 
 </ruleset>
