+++ pmd-xml/src/main/resources/rulesets/xml/rulesets.properties
@@ -2,4 +2,17 @@
 # BSD-style license; for more info see http://pmd.sourceforge.net/license.html
 #
 
-rulesets.filenames=rulesets/xml/basic.xml
+rulesets.filenames=\
+    category/xml/errorprone.xml
+
+#
+# categories without rules
+#
+# category/xml/bestpractices.xml
+# category/xml/codestyle.xml
+# category/xml/design.xml
+# category/xml/documentation.xml
+# category/xml/multithreading.xml
+# category/xml/performance.xml
+# category/xml/security.xml
+
