+++ pmd-xml/src/main/resources/rulesets/xsl/rulesets.properties
@@ -2,4 +2,17 @@
 # BSD-style license; for more info see http://pmd.sourceforge.net/license.html
 #
 
-rulesets.filenames=rulesets/xsl/xpath.xml
+rulesets.filenames=\
+    category/xsl/codestyle.xml,\
+    category/xsl/performance.xml
+
+#
+# categories without rules
+#
+# category/xsl/bestpractices.xml
+# category/xsl/design.xml
+# category/xsl/documentation.xml
+# category/xsl/errorprone.xml
+# category/xsl/multithreading.xml
+# category/xsl/security.xml
+
