+++ pmd-xml/src/test/java/net/sourceforge/pmd/lang/pom/rule/codestyle/CodeStyleRulesTest.java
@@ -2,17 +2,16 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.pom.rule.basic;
+package net.sourceforge.pmd.lang.pom.rule.codestyle;
 
 import net.sourceforge.pmd.testframework.SimpleAggregatorTst;
 
-public class BasicRulesTest extends SimpleAggregatorTst {
+public class CodeStyleRulesTest extends SimpleAggregatorTst {
 
-    private static final String RULESET = "pom-basic";
+    private static final String RULESET = "category/pom/codestyle.xml";
 
     @Override
     public void setUp() {
         addRule(RULESET, "ProjectVersionAsDependencyVersion");
-        addRule(RULESET, "InvalidDependencyTypes");
     }
 }
