+++ pmd-xml/src/test/java/net/sourceforge/pmd/lang/xml/rule/errorprone/ErrorProneRulesTest.java
@@ -2,13 +2,13 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.xml.rule.basic;
+package net.sourceforge.pmd.lang.xml.rule.errorprone;
 
 import net.sourceforge.pmd.testframework.SimpleAggregatorTst;
 
-public class BasicRulesTest extends SimpleAggregatorTst {
+public class ErrorProneRulesTest extends SimpleAggregatorTst {
 
-    private static final String RULESET = "xml-basic";
+    private static final String RULESET = "category/xml/errorprone.xml";
 
     @Override
     public void setUp() {
