+++ pmd-xml/src/test/resources/net/sourceforge/pmd/lang/pom/rule/codestyle/xml/ProjectVersionAsDependencyVersion.xml
