+++ pmd-xml/src/test/resources/net/sourceforge/pmd/lang/pom/rule/errorprone/xml/InvalidDependencyTypes.xml
