+++ pmd-xml/src/test/resources/net/sourceforge/pmd/lang/xsl/rule/codestyle/xml/UseConcatOnce.xml
