+++ pmd-xml/src/test/resources/net/sourceforge/pmd/lang/xsl/rule/performance/xml/AvoidAxisNavigation.xml
